Fix time overflow

+++ source/main.cpp
@@ -9,9 +9,9 @@
 TemporalGraph * build(char * argv[]) {
 
     std::cout << "Building graph..." << std::endl;
-    int build_graph_start_time = currentTime();
+    unsigned long long build_graph_start_time = currentTime();
     TemporalGraph * Graph = new TemporalGraph(argv[1], (char *)"Undirected");
-    int build_graph_end_time = currentTime();
+    unsigned long long build_graph_end_time = currentTime();
     std::cout << "Build graph success in " << timeFormatting(difftime(build_graph_end_time, build_graph_start_time)).str() << std::endl;
     std::cout << "n = " << Graph->numOfVertices() << ", m = " << Graph->numOfEdges() << ", tmax = " << Graph->tmax << std::endl;
     return Graph;
@@ -22,7 +22,7 @@ int main(int argc, char * argv[]) {
 
     std::ios::sync_with_stdio(false);
 
-    int start_time = currentTime();
+    unsigned long long start_time = currentTime();
 
     if (argc != 5) {
         std::cout << "Parameter error. Please check the readme file." << std::endl;
@@ -35,9 +35,9 @@ int main(int argc, char * argv[]) {
 
     if (std::strcmp(argv[4], "Online") == 0) {
         std::cout << "Running online search..." << std::endl;
-        int online_search_start_time = currentTime();
+        unsigned long long online_search_start_time = currentTime();
         online(Graph, argv[2], argv[3]);
-        int online_search_end_time = currentTime();
+        unsigned long long online_search_end_time = currentTime();
         std::cout << "Online search completed in " << timeFormatting(difftime(online_search_end_time, online_search_start_time)).str() << std::endl;
         delete Graph;
     }
@@ -45,16 +45,16 @@ int main(int argc, char * argv[]) {
     if (std::strcmp(argv[4], "Baseline") == 0) {
         std::cout << "Running baseline..." << std::endl;
         std::cout << "Constructing the index structure..." << std::endl;
-        int index_construction_start_time = currentTime();
+        unsigned long long index_construction_start_time = currentTime();
         BaselineIndex *Index = new BaselineIndex(Graph);
-        int index_construction_end_time = currentTime();
+        unsigned long long index_construction_end_time = currentTime();
         std::cout << "Index construction completed in " << timeFormatting(difftime(index_construction_end_time, index_construction_start_time)).str() << std::endl;
         std::cout << "Index cost " << Index->size() << " bytes" << std::endl;
         delete Graph;
         std::cout << "Solving queries..." << std::endl;
-        int query_start_time = currentTime();
+        unsigned long long query_start_time = currentTime();
         baseline(Index, vertex_num, argv[2], argv[3]);
-        int query_end_time = currentTime();
+        unsigned long long query_end_time = currentTime();
         std::cout << "Query completed in " << timeFormatting(difftime(query_end_time, query_start_time)).str() << std::endl;
         std::cout << "Baseline completed!" << std::endl;
     }
@@ -62,15 +62,15 @@ int main(int argc, char * argv[]) {
     if (std::strcmp(argv[4], "OBaseline") == 0) {
         std::cout << "Running optimized baseline..." << std::endl;
         std::cout << "Constructing the index structure..." << std::endl;
-        int index_construction_start_time = currentTime();
+        unsigned long long index_construction_start_time = currentTime();
         OptimizedBaseline *Index = new OptimizedBaseline(Graph);
-        int index_construction_end_time = currentTime();
+        unsigned long long index_construction_end_time = currentTime();
         std::cout << "Index construction completed in " << timeFormatting(difftime(index_construction_end_time, index_construction_start_time)).str() << std::endl;
         delete Graph;
         std::cout << "Solving queries..." << std::endl;
-        int query_start_time = currentTime();
+        unsigned long long query_start_time = currentTime();
         optimized_baseline(Index, vertex_num, argv[2], argv[3]);
-        int query_end_time = currentTime();
+        unsigned long long query_end_time = currentTime();
         std::cout << "Query completed in " << timeFormatting(difftime(query_end_time, query_start_time)).str() << std::endl;
         std::cout << "Optimized Baseline completed!" << std::endl;
     }
@@ -78,15 +78,15 @@ int main(int argc, char * argv[]) {
     if (std::strcmp(argv[4], "Kruskal") == 0) {
         std::cout << "Running kruskal reconstruction tree..." << std::endl;
         std::cout << "Constructing the index structure..." << std::endl;
-        int index_construction_start_time = currentTime();
+        unsigned long long index_construction_start_time = currentTime();
         KruskalReconstructionTree *Index = new KruskalReconstructionTree(Graph);
-        int index_construction_end_time = currentTime();
+        unsigned long long index_construction_end_time = currentTime();
         std::cout << "Index construction completed in " << timeFormatting(difftime(index_construction_end_time, index_construction_start_time)).str() << std::endl;
         delete Graph;
         std::cout << "Solving queries..." << std::endl;
-        int query_start_time = currentTime();
+        unsigned long long query_start_time = currentTime();
         kruskal(Index, vertex_num, argv[2], argv[3]);
-        int query_end_time = currentTime();
+        unsigned long long query_end_time = currentTime();
         std::cout << "Query completed in " << timeFormatting(difftime(query_end_time, query_start_time)).str() << std::endl;
         std::cout << "Kruskal reconstruction tree completed!" << std::endl;
     }
@@ -94,21 +94,21 @@ int main(int argc, char * argv[]) {
     if (std::strcmp(argv[4], "DKruskal") == 0) {
         std::cout << "Running differential kruskal reconstruction tree..." << std::endl;
         std::cout << "Constructing the index structure..." << std::endl;
-        int index_construction_start_time = currentTime();
+        unsigned long long index_construction_start_time = currentTime();
         DifferentialKruskal *Index = new DifferentialKruskal(Graph);
-        int index_construction_end_time = currentTime();
+        unsigned long long index_construction_end_time = currentTime();
         std::cout << "Index construction completed in " << timeFormatting(difftime(index_construction_end_time, index_construction_start_time)).str() << std::endl;
         std::cout << "Index cost " << Index->size() << " bytes" << std::endl;
         delete Graph;
         std::cout << "Solving queries..." << std::endl;
-        int query_start_time = currentTime();
+        unsigned long long query_start_time = currentTime();
         differential_kruskal(Index, vertex_num, argv[2], argv[3]);
-        int query_end_time = currentTime();
+        unsigned long long query_end_time = currentTime();
         std::cout << "Query completed in " << timeFormatting(difftime(query_end_time, query_start_time)).str() << std::endl;
         std::cout << "Differential kruskal reconstruction tree completed!" << std::endl;
     }
 
-    int end_time = currentTime();
+    unsigned long long end_time = currentTime();
     std::cout << "Program finished in " << timeFormatting(difftime(end_time, start_time)).str() << std::endl;
 
     return 0;
